Split numbered from Iterable
